[cosmetics] Drop useless `StateActuator` interface
This commit moves the `state_actuator.go` content into the `state_reconciler.go` file for simplicity reasons. There is no need to abstract the `StateActuator` interface.
